Move critical editor style into JS
Summary:There have been a number of bug reports caused by Draft.css not being included. The main one is that the editor can start flipping out if `white-space: pre-wrap` is not applied.

Users should generally be including Draft.css, and should be aware of behavior gaps if they choose not to. Documenting this accordingly.
Closes https://github.com/facebook/draft-js/pull/159

Reviewed By: spicyj

Differential Revision: D3013416

fb-gh-sync-id: 6423b84be619fa1013ef7babba30be8956456a97
shipit-source-id: 6423b84be619fa1013ef7babba30be8956456a97