Documentation updates
Summary:Added some missing documentation.

I didn't really document `RichUtils`, I just added a placeholder so it's at least easy to find out which methods are available without having to resort to the source code (and the names should be self explanatory enough for it to be good for now).

Resolves #263
Closes https://github.com/facebook/draft-js/pull/269

Differential Revision: D3139568

fb-gh-sync-id: 34cfa671c189b561c6a84c5ba991be224019dbe1
fbshipit-source-id: 34cfa671c189b561c6a84c5ba991be224019dbe1