Switch to the more conservative Immutable.js ~3.7.4 version to avoid upstream flow definition errors.
Summary:
The `3.8.x` versions of Immutable.js have changed `#toOrderedMap(): OrderedMap` to `#toOrderedMap(): Map` which breaks Draft.js — so I propose we use `~3.7.4` instead of `^3.7.4`.

Draft.js's definition of `BlockMap` is `OrderedMap<string, ContentBlock>` which is not compatible with `Map`.

In the interim (until Immutable.js makes a decision as to whether `#toOrderedMap(): Map` is correct) using the `3.7.x` range allows Draft.js `master` to pass flow checks.
Closes https://github.com/facebook/draft-js/pull/468

Differential Revision: D3456991

fbshipit-source-id: 0e9a176d800d3216d30b00ae1da1a682ee93273b