Part two deprecate but dont delete old entity api (#814)
* Add warnings and comments about deprecating old Entity API

We are deprecating the old DraftEntity API.

Why? Because
 - currently changing or adding entities won't trigger a re-render
 - the use of a non-Immutable storage for entities doesn't fit well with
   the rest of the Draft architecture
 - moves Entity storage out of global module; this just makes more sense
   as an API, rather than having it globally available.

This moves the logic for the old API into private methods, and adds
warnings and comments around the public-but-deprecated API methods.

We then use the private methods internally while we still support both
APIs.

A future version of Draft will actually remove the old API completely,
and these warnings and comments will help folks make updates to their
code to prepare.

We will release more documentation about this soon.

* Update syntax of tex example

This example was still using the old syntax, and the new warnings we
added alerted us to this fact when doing manual testing.