Fix website's rich text editor example
Summary:
**Summary**
PR #1346 broke the example as it didn't update all the code needed for handleKeyCommand to correctly pass the editorState.

This change mirrors the same change applied to rich.html.

**Test Plan**
Go to the website's main page `/index.html`, write text on the editor and try to delete it, it should work as expected.
Closes https://github.com/facebook/draft-js/pull/1375

Differential Revision: D5800218

fbshipit-source-id: 3b24afcda75ead7c2854cd1236a3fba956c34768