Exploration - Feature tree / Wiring up paste processor
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Wiring up paste processor**

This wires up the paste processor to make sure it pass the tree flag along. We will revisit this approach on a follow up diff where we will change the flag to actually be a GK feature flag instead of function feature argument

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1523

Differential Revision: D6382429

fbshipit-source-id: 950fc0a87b7e61f83bd134a2811947a2063396a4