Exploration - Feature tree / Fixing shouldComponentUpdate lifcycle for ContentBlockNode
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Fixing component lifcycle for ContentBlockNode**

`block.getChildKeys` is not a property it is meant to be a function call
***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1525

Differential Revision: D6389650

fbshipit-source-id: d9923b2042beed997cd63f83c7beb4853496be87